agentic-flow/pyproject.toml
[project]
name = "rag-pjr"
version = "0.1.0"
description = ""
authors = [
    {name = "Ekansh Jain",email = "ekansh.jain@fiftyfivetech.io"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "aiohappyeyeballs (==2.4.6)",
    "aiohttp (==3.11.12)",
    "aiosignal (==1.3.2)",
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "astrapy (==1.5.2)",
    "async-timeout (==4.0.3)",
    "attrs (==25.1.0)",
    "cachetools (==5.5.2)",
    "cassandra-driver (==3.29.2)",
    "certifi (==2025.1.31)",
    "charset-normalizer (==3.4.1)",
    "click (==8.1.8)",
    "dataclasses-json (==0.6.7)",
    "deprecation (==2.1.0)",
    "dnspython (==2.7.0)",
    "docstring-parser (==0.16)",
    "exceptiongroup (==1.2.2)",
    "fastapi (==0.115.11)",
    "filelock (==3.17.0)",
    "frozenlist (==1.5.0)",
    "fsspec (==2025.2.0)",
    "geomet (==0.2.1.post1)",
    "google-ai-generativelanguage (==0.6.15)",
    "google-api-core (==2.24.1)",
    "google-api-python-client (==2.161.0)",
    "google-auth (==2.38.0)",
    "google-auth-httplib2 (==0.2.0)",
    "google-cloud-aiplatform (==1.81.0)",
    "google-cloud-bigquery (==3.29.0)",
    "google-cloud-core (==2.4.2)",
    "google-cloud-resource-manager (==1.14.1)",
    "google-cloud-storage (==2.19.0)",
    "google-crc32c (==1.6.0)",
    "google-generativeai (==0.8.4)",
    "google-resumable-media (==2.7.2)",
    "googleapis-common-protos (==1.68.0)",
    "greenlet (==3.1.1)",
    "grpc-google-iam-v1 (==0.14.0)",
    "grpcio (==1.70.0)",
    "grpcio-status (==1.70.0)",
    "h11 (==0.14.0)",
    "h2 (==4.2.0)",
    "hpack (==4.1.0)",
    "httpcore (==1.0.7)",
    "httplib2 (==0.22.0)",
    "httpx (==0.28.1)",
    "httpx-sse (==0.4.0)",
    "huggingface-hub (==0.29.1)",
    "hyperframe (==6.1.0)",
    "idna (==3.10)",
    "jinja2 (==3.1.5)",
    "joblib (==1.4.2)",
    "jsonpatch (==1.33)",
    "jsonpointer (==3.0.0)",
    "langchain (==0.3.19)",
    "langchain-community (==0.3.18)",
    "langchain-core (==0.3.40)",
    "langchain-google-vertexai (==2.0.14)",
    "langchain-huggingface (==0.1.2)",
    "langchain-text-splitters (==0.3.6)",
    "langsmith (==0.3.11)",
    "markupsafe (==3.0.2)",
    "marshmallow (==3.26.1)",
    "mpmath (==1.3.0)",
    "multidict (==6.1.0)",
    "mypy-extensions (==1.0.0)",
    "networkx (==3.4.2)",
    "numpy (==1.26.4)",
    "nvidia-cublas-cu12 (==12.4.5.8)",
    "nvidia-cuda-cupti-cu12 (==12.4.127)",
    "nvidia-cuda-nvrtc-cu12 (==12.4.127)",
    "nvidia-cuda-runtime-cu12 (==12.4.127)",
    "nvidia-cudnn-cu12 (==9.1.0.70)",
    "nvidia-cufft-cu12 (==11.2.1.3)",
    "nvidia-curand-cu12 (==10.3.5.147)",
    "nvidia-cusolver-cu12 (==11.6.1.9)",
    "nvidia-cusparse-cu12 (==12.3.1.170)",
    "nvidia-cusparselt-cu12 (==0.6.2)",
    "nvidia-nccl-cu12 (==2.21.5)",
    "nvidia-nvjitlink-cu12 (==12.4.127)",
    "nvidia-nvtx-cu12 (==12.4.127)",
    "orjson (==3.10.15)",
    "packaging (==23.2)",
    "pandas (==2.2.3)",
    "pillow (==11.1.0)",
    "propcache (==0.3.0)",
    "proto-plus (==1.26.0)",
    "protobuf (==5.29.3)",
    "psycopg2 (==2.9.10)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.1)",
    "pydantic (==2.10.6)",
    "pydantic-settings (==2.8.0)",
    "pydantic-core (==2.27.2)",
    "pymongo (==4.11.1)",
    "pymupdf (==1.25.3)",
    "pyparsing (==3.2.1)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "pytz (==2025.1)",
    "pyyaml (==6.0.2)",
    "regex (==2024.11.6)",
    "requests (==2.32.3)",
    "requests-toolbelt (==1.0.0)",
    "rsa (==4.9)",
    "safetensors (==0.5.2)",
    "scikit-learn (==1.6.1)",
    "scipy (==1.15.2)",
    "sentence-transformers (==3.4.1)",
    "shapely (==2.0.7)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.38)",
    "starlette (==0.46.0)",
    "sympy (==1.13.1)",
    "tenacity (==8.5.0)",
    "threadpoolctl (==3.5.0)",
    "tokenizers (==0.21.0)",
    "toml (==0.10.2)",
    "torch (==2.6.0)",
    "tqdm (==4.67.1)",
    "transformers (==4.49.0)",
    "triton (==3.2.0)",
    "typing-inspect (==0.9.0)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2025.1)",
    "uritemplate (==4.1.1)",
    "urllib3 (==2.3.0)",
    "uuid6 (==2024.7.10)",
    "yarl (==1.18.3)",
    "zstandard (==0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
